requirements.txt
streamlit
pdfplumber
python-docx
pandas